Fix a forgotten cast that made building fail
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@11090 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/animations/animation_base.cpp
@@ -135,7 +135,7 @@ void AnimationBase::computeLengths()
     for_in (curr, m_all_ipos)
     {
         const std::vector<core::vector2df>& points = curr->getPoints();
-        max_points = std::max(max_points, points.size());
+        max_points = std::max(max_points, (unsigned int) points.size());
     }
 
     // Divide (on average) each segment into STEPS points, and use
